pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pharmvar_tools"
version = "0.4.0"
authors = [
  { name="Mark Santcroos", email="m.a.santcroos@lumc.nl" },
  { name="Jonathan K. Vis", email="j.k.vis@lumc.nl" },
]
description = "Python tools to work with Pharmvar data."
readme = "README.md"
requires-python = ">=3.7"
dependencies = ["mutalyzer-algebra", "networkx", "ratelimiter", "requests", "pyyaml"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
pharmvar-check = "pharmvar_tools.check:main"
pharmvar-compare = "pharmvar_tools.compare:main"
pharmvar-plot = "pharmvar_tools.to_dot:main"
pharmvar-lcsgraphs = "pharmvar_tools.lcsgraphs:main"


[project.urls]
"Homepage" = "https://github.com/lumc-pgx/pharmvar-tools"
"Bug Tracker" = "https://github.com/lumc-pgx/pharmvar-tools/issues"


[tool.setuptools.package-data]
pharmvar_tools = ["config/*.yaml"]


[project.optional-dependencies]
dev = ["build", "twine"]
